Ver 0.2

+++ src/main/java/hr/klobucaric/webshop/address/AddressService.java
@@ -4,5 +4,5 @@ package hr.klobucaric.webshop.address;
 public interface AddressService {
 
     AddressDto createAddress(AddressCommand addressCommand);
-    void setDefaultAddress(Integer id);
+    void setDefaultAddress(Long id);
 }
